style: removes all short-form identifiers in favor of fully formed URIs
This is in line with JSON-LD 1.1 best practices and there were some lingering changes from the old suites that were being kept identical to the VC context. However, because the proofPurpose change makes that no longer possible we should be updating this to no longer use short form identifiers instead opting for the the fully formed URIs.